servers/mcp_playwright_server/pyproject.toml
[project]
name = "mcp-playwright-server"
version = "0.1.0"
description = "Playwright server for MCP tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.4.1",
    "pydantic>=2.10.6",
    "pytest-playwright>=0.7.0",
]
packages = [
    { include = "mcp_playwright_server", from = "src" }
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_playwright_server"]